finish project

+++ src/controllers/post-controller.js
@@ -1,5 +1,5 @@
 const fs = require("fs");
-const { Post } = require("../models");
+const { Post, User } = require("../models");
 const { validateCreatePost } = require("../validators/post-validator");
 
 exports.createPost = async (req, res, next) => {
@@ -9,10 +9,19 @@ exports.createPost = async (req, res, next) => {
       title: req.body.title,
     });
 
-    value.userId = req.user.id;
+    value.userId = req.user.id; // #userId in post
 
     //#insert data to db
-    const post = await Post.create(value);
+    const createdPost = await Post.create(value);
+    const post = await Post.findOne({
+      where: {
+        id: +createdPost.id,
+      },
+      include: {
+        model: User,
+      },
+    });
+
     res.status(201).json(post);
   } catch (err) {
     next(err);
@@ -23,9 +32,42 @@ exports.getAllPost = async (req, res, next) => {
   try {
     const posts = await Post.findAll({
       where: { groupId: req.params.groupId },
+      order: [["createdAt", "DESC"]],
+      include: [
+        {
+          model: User,
+
+          attributes: {
+            exclude: ["password"],
+          },
+        },
+      ],
     });
 
     res.status(200).json(posts);
+  } catch (err) {
+    next(err);
+  }
+};
+
+exports.editPost = async (req, res, next) => {
+  try {
+    const value = req.body;
+    console.log("the value sent is: ", req.body);
+    console.log("The post id is : ", req.params);
+    const checkPost = await Post.findOne({ where: { id: +req.params.postId } });
+    // const newPost = await Post.findOne({
+    //   where: { id: req.params },
+    // });
+
+    if (!checkPost) {
+      console.log("The post id does not exist");
+    }
+    const newPost = await Post.update(value, {
+      where: { id: +req.params.postId, userId: req.user.id },
+      // userId: req.user.id = only this id allowed to edit this post
+    });
+    res.status(200).json({ message: "Sucessful edit post", checkPost });
   } catch (err) {
     next(err);
   }
@@ -35,16 +77,7 @@ exports.deletePost = async (req, res, next) => {
   const { postId } = req.params;
 
   try {
-    // const post = await Post.findOne({
-    //   where: { groupId: postId },
-    // });
-    // if (!post) {
-    //   createError("this post was not found", 400);
-    // }
-    // if (post.userId !== req.user.id) {
-    //   createError("You are unauthorized", 403);
-    // }
-    await Post.destroy({ where: { id: postId } });
+    await Post.destroy({ where: { id: postId, userId: req.user.id } });
     res.status(204).json();
   } catch (err) {
     next(err);
